+++ CHANGELOG.md
@@ -7,6 +7,7 @@ changes with their rationale when appropriate:
 - **http4k-*** : Upgrade some dependency versions.
 - **http4k-connect-amazon-s3*** :Added support for s3 bucket keys containing hash (#). H/T @aharin
 - **http4k-mcp-sdk**: Updates to support sampling callbacks in Tool and other requests.
+- **http4k-mcp-sdk**: Add filters for all common MCP Request types
 
 ### v6.5.1.0
 - **http4k-mcp-sdk**: Tidy up completions API.